backend: Move Rhyolite.Backend.Logging to its own package and re-export
